move updateNotifier to node.js, delete update.js
