+++ package.json
@@ -54,7 +54,6 @@
   },
   "browser": {
     "./node.js": "./browser.js",
-    "./update.js": false,
     "path": false
   },
   "pre-commit": [
